added project
Project contains two classes. Need to enter the input text file location manually in both the classes. An input text file is already provided in the java project. Get it's file location & replace with the existing file location used in the both the classes.

then the program should work perfectly.